update like/unlike feedback

+++ BusinessObject/DTOs/Response/Courses/CourseDetailWithoutProgressResponse.cs
@@ -5,7 +5,7 @@ namespace BusinessObject.DTOs.Response.Courses
 {
     public class CourseDetailWithoutProgressResponse
     {
-        public long CourseId { get; set; }
+        public int CourseId { get; set; }
         public string Name { get; set; } = string.Empty;
         public string Description { get; set; } = string.Empty;
         public string Status { get; set; } = string.Empty;
@@ -19,7 +19,7 @@ namespace BusinessObject.DTOs.Response.Courses
 
     public class CourseSectionDetailWithoutProgressResponse
     {
-        public long CourseSectionId { get; set; }
+        public int CourseSectionId { get; set; }
         public string Name { get; set; } = string.Empty;
         public string? Description { get; set; }
         public int Order { get; set; }
@@ -29,7 +29,7 @@ namespace BusinessObject.DTOs.Response.Courses
 
     public class LessonDetailWithoutProgressResponse
     {
-        public long LessonId { get; set; }
+        public int LessonId { get; set; }
         public string Name { get; set; } = string.Empty;
         public string? Description { get; set; }
         public int Order { get; set; }
@@ -41,7 +41,7 @@ namespace BusinessObject.DTOs.Response.Courses
 
     public class LessonItemDetailWithoutProgressResponse
     {
-        public long LessonItemId { get; set; }
+        public int LessonItemId { get; set; }
         public string Name { get; set; } = string.Empty;
         public int Order { get; set; }
     }
