+++ Repository/IRepositories/ICoursesectionRepository.cs
@@ -5,8 +5,8 @@ namespace Repository.IRepositories
 {
     public interface ICoursesectionRepository : IBaseRepository<Coursesection>
     {
-        Task<IEnumerable<Coursesection>> GetByCourseIdAsync(long courseId);
-        Task<Coursesection?> GetByCourseSectionIdAsync(long courseSectionId);
-        Task<IEnumerable<Coursesection>> GetActiveSectionsByCourseAsync(long courseId);
+        Task<IEnumerable<Coursesection>> GetByCourseIdAsync(int courseId);
+        Task<Coursesection?> GetByCourseSectionIdAsync(int courseSectionId);
+        Task<IEnumerable<Coursesection>> GetActiveSectionsByCourseAsync(int courseId);
     }
 }
