+++ Repository/IRepositories/IProcessRepository.cs
@@ -5,13 +5,12 @@ namespace Repository.IRepositories
 {
     public interface IProcessRepository : IBaseRepository<Process>
     {
-        Task<IEnumerable<Process>> GetByLessonAsync(long lessonId);
-        // New student-based APIs (no courseKey)
-        Task<Process?> GetByStudentAndLessonAsync(string studentId, long lessonId);
-        Task<IEnumerable<Process>> GetByStudentAndCourseAsync(string studentId, long courseId);
-        Task<bool> IsLessonCompletedForStudentAsync(string studentId, long lessonId);
+        Task<IEnumerable<Process>> GetByLessonAsync(int lessonId);
+        Task<Process?> GetByStudentAndLessonAsync(string studentId, int lessonId);
+        Task<IEnumerable<Process>> GetByStudentAndCourseAsync(string studentId, int courseId);
+        Task<bool> IsLessonCompletedForStudentAsync(string studentId, int lessonId);
         Task<IEnumerable<Process>> GetByStudentAsync(string studentId);
         Task<Process> UnlockLessonForStudentAsync(string studentId, int lessonId, int courseId);
-        Task<bool> HasStudentCompletedCourseAsync(string studentId, long courseId);
+        Task<bool> HasStudentCompletedCourseAsync(string studentId, int courseId);
     }
 }
