+++ Repository/IRepositories/IProcessitemRepository.cs
@@ -5,8 +5,8 @@ namespace Repository.IRepositories
 {
     public interface IProcessitemRepository : IBaseRepository<Processitem>
     {
-        Task<IEnumerable<Processitem>> GetByProcessIdAsync(long processId);
-        Task<Processitem?> GetByProcessAndLessonItemAsync(long processId, long lessonItemId);
-        Task<bool> IsLessonItemCompletedAsync(long processId, long lessonItemId);
+        Task<IEnumerable<Processitem>> GetByProcessIdAsync(int processId);
+        Task<Processitem?> GetByProcessAndLessonItemAsync(int processId, int lessonItemId);
+        Task<bool> IsLessonItemCompletedAsync(int processId, int lessonItemId);
     }
 }
