+++ Repository/Repositories/ProcessitemRepository.cs
@@ -15,7 +15,7 @@ namespace Repository.Repositories
             _context = context;
         }
 
-        public async Task<IEnumerable<Processitem>> GetByProcessIdAsync(long processId)
+        public async Task<IEnumerable<Processitem>> GetByProcessIdAsync(int processId)
         {
             return await _context.Set<Processitem>()
                 .Include(pi => pi.LessonItem)
@@ -24,7 +24,7 @@ namespace Repository.Repositories
                 .ToListAsync();
         }
 
-        public async Task<Processitem?> GetByProcessAndLessonItemAsync(long processId, long lessonItemId)
+        public async Task<Processitem?> GetByProcessAndLessonItemAsync(int processId, int lessonItemId)
         {
             return await _context.Set<Processitem>()
                 .Include(pi => pi.LessonItem)
@@ -32,7 +32,7 @@ namespace Repository.Repositories
                 .FirstOrDefaultAsync(pi => pi.ProcessId == processId && pi.LessonItemId == lessonItemId);
         }
 
-        public async Task<bool> IsLessonItemCompletedAsync(long processId, long lessonItemId)
+        public async Task<bool> IsLessonItemCompletedAsync(int processId, int lessonItemId)
         {
             var processItem = await GetByProcessAndLessonItemAsync(processId, lessonItemId);
             return processItem != null;
